refactor(tui): split small layout responsibilities
