+++ packages/bijou-tui/src/layout-preset.ts
@@ -142,36 +142,8 @@ export function restoreLayoutState(
     splitRatiosByPage,
   };
 }
-
-/** Preset: side-by-side split with equal ratio. */
-export function presetSideBySide(splitId: string): SerializedPageLayout {
-  return {
-    splitRatios: { [splitId]: 0.5 },
-    focusedPane: undefined,
-    minimized: [],
-    dockOrder: {},
-    maximizedPane: undefined,
-  };
-}
-
-/** Preset: stacked (vertical) split with equal ratio. */
-export function presetStacked(splitId: string): SerializedPageLayout {
-  return {
-    splitRatios: { [splitId]: 0.5 },
-    focusedPane: undefined,
-    minimized: [],
-    dockOrder: {},
-    maximizedPane: undefined,
-  };
-}
-
-/** Preset: focused on a single pane (maximized). */
-export function presetFocused(paneId: string): SerializedPageLayout {
-  return {
-    splitRatios: {},
-    focusedPane: paneId,
-    minimized: [],
-    dockOrder: {},
-    maximizedPane: paneId,
-  };
-}
+export {
+  presetFocused,
+  presetSideBySide,
+  presetStacked,
+} from './layout-preset-factories.js';
