+++ packages/bijou-tui/src/layout.ts
@@ -114,64 +114,4 @@ export function place(content: string, options: PlaceOptions): string {
   return result.join('\n');
 }
 
-// ---------------------------------------------------------------------------
-// vstack / hstack
-// ---------------------------------------------------------------------------
-
-/**
- * Vertical stack — join blocks with newlines.
- *
- * @param blocks - One or more rendered string blocks to stack vertically.
- * @returns A single string with blocks separated by newlines.
- */
-export function vstack(...blocks: string[]): string {
-  return blocks.join('\n');
-}
-
-/**
- * Horizontal stack — place blocks side by side with a gap between columns.
- * Pads shorter blocks with empty lines to align rows.
- *
- * @param gap    - Number of space characters between adjacent columns.
- * @param blocks - One or more rendered string blocks to stack horizontally.
- * @returns A single string with blocks arranged side by side.
- */
-export function hstack(gap: number, ...blocks: string[]): string {
-  if (blocks.length === 0) return '';
-  if (blocks.length === 1) return blocks[0] ?? '';
-
-  const split = blocks.map((b) => b.split('\n'));
-  const maxRows = Math.max(...split.map((lines) => lines.length));
-  const widths = split.map((lines) => Math.max(...lines.map(visualWidth)));
-  const spacer = ' '.repeat(Math.max(0, gap));
-
-  const rows: string[] = [];
-  for (let r = 0; r < maxRows; r++) {
-    const parts: string[] = [];
-    for (let c = 0; c < split.length; c++) {
-      const lines = split[c];
-      const line = lines?.[r] ?? '';
-      // Last column doesn't need right-padding
-      if (c < split.length - 1) {
-        const width = widths[c] ?? 0;
-        parts.push(line + ' '.repeat(Math.max(0, width - visualWidth(line))));
-      } else {
-        parts.push(line);
-      }
-    }
-    rows.push(parts.join(spacer).trimEnd());
-  }
-  return rows.join('\n');
-}
-
-/**
- * Compute the visible width of a string after stripping ANSI escape sequences.
- *
- * @param s - The string possibly containing ANSI color/style codes.
- * @returns The character count excluding escape sequences.
- */
-function visualWidth(s: string): number {
-  // Strip ANSI escape sequences
-  // eslint-disable-next-line no-control-regex
-  return s.replace(/\x1b\[[0-9;]*m/g, '').length;
-}
+export { hstack, vstack } from './layout-stack.js';
