+++ packages/bijou-tui/src/subapp/mount.ts
@@ -37,10 +37,6 @@ export interface MountedApp<SubMsg> {
   cmds: Cmd<SubMsg>[];
 }
 
-export type SubAppAdapterCases<ParentMsg, SubMsg extends { readonly type: string }> = {
-  readonly [K in SubMsg['type']]: (msg: Extract<SubMsg, { type: K }>) => ParentMsg;
-};
-
 /**
  * Mount a sub-app, returning its rendered surface and any pending commands.
  *
@@ -61,43 +57,6 @@ export function mount<SubModel, SubMsg extends object, ParentMsg>(
   const surfaceOrNode = app.view(model);
 
   return [surfaceOrNode, []];
-}
-
-/**
- * Build an exhaustive discriminant-based sub-app message mapper.
- *
- * The returned function matches the `onMsg` signature expected by
- * `initSubApp()`, `updateSubApp()`, and `mount()`, while TypeScript enforces
- * that every `SubMsg['type']` variant is covered up front.
- */
-export function createSubAppAdapter<ParentMsg, SubMsg extends { readonly type: string }>(
-  cases: SubAppAdapterCases<ParentMsg, SubMsg>,
-): (msg: SubMsg) => ParentMsg {
-  return (msg) => {
-    if (hasHandler(cases, msg)) {
-      const handler = handlerFor(cases, msg.type);
-      return handler(msg);
-    }
-    throw new Error(`Unhandled sub-app message type: ${msg.type}`);
-  };
-}
-
-function handlerFor<
-  ParentMsg,
-  SubMsg extends { readonly type: string },
-  Type extends SubMsg['type'],
->(
-  cases: SubAppAdapterCases<ParentMsg, SubMsg>,
-  type: Type,
-): (msg: Extract<SubMsg, { type: Type }>) => ParentMsg {
-  return cases[type];
-}
-
-function hasHandler<ParentMsg, SubMsg extends { readonly type: string }>(
-  cases: SubAppAdapterCases<ParentMsg, SubMsg>,
-  msg: SubMsg,
-): msg is Extract<SubMsg, { readonly type: string }> {
-  return Object.hasOwn(cases, msg.type);
 }
 
 export interface SubAppOptions<SubMsg extends object, ParentMsg> {
@@ -177,3 +136,8 @@ function isSubAppMessageResult<SubMsg extends object>(
 ): result is SubMsg {
   return result !== undefined && result !== QUIT && !isCmdCleanup(result);
 }
+
+export {
+  createSubAppAdapter,
+  type SubAppAdapterCases,
+} from './adapter.js';
